Move itoa's static numlen into public library

+++ ft_strdup.c
@@ -6,16 +6,17 @@
 /*   By: jgelbard <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/02/20 15:24:18 by jgelbard          #+#    #+#             */
-/*   Updated: 2018/03/08 21:17:44 by jgelbard         ###   ########.fr       */
+/*   Updated: 2018/03/23 19:00:39 by jgelbard         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_string.h"
+#include <string.h>
 #include <stdlib.h>
 
 char	*ft_strdup(const char *s1)
 {
-	int		n;
+	size_t	n;
 	char	*s2;
 
 	n = ft_strlen(s1);
